requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.5.30
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
Django==3.2.7
django-bootstrap4==3.0.1
django-forms-bootstrap==3.1.0
django-registration==3.2
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
